# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "2.0.0",
    "title": "DREAM Lab API",
    "description": "Endpoints para la API de DREAM Lab. Puede encontrar más información en [GitHub](https://github.com/CodeCraft-Solutions-DREAM-Lab).",
    "contact": {
      "name": "CodeCraft Solutions",
      "url": "https://github.com/CodeCraft-Solutions-DREAM-Lab"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local server"
    },
    {
      "url": "https://dreamlab-api.azurewebsites.net/",
      "description": "Live server"
    }
  ],
  "paths": {
    "/auth/usuario": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica un usuario",
        "description": "Autentica un usuario",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jwt": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c3VhcmlvIjoiYTAxMTc3NzY3IiwiaWF0IjoxNzE1NzUzNzQzfQ.ml-vMvWq5X8_FdILT9YIPv0oPc9Wlvj3f_N4VhHCAZA"
                    },
                    "rol": {
                      "type": "string",
                      "example": "Regular"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string",
                    "example": "test"
                  },
                  "contrasena": {
                    "type": "string",
                    "example": "test"
                  },
                  "origen": {
                    "type": "string",
                    "example": "qr"
                  },
                  "tagId": {
                    "type": "string",
                    "example": "test"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verifica un token",
        "description": "Verifica un token",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isAuth": {
                      "type": "boolean",
                      "example": true
                    },
                    "token_data": {
                      "type": "object",
                      "properties": {
                        "datosUsuario": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isAuth": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c3VhcmlvIjoiYTAxMTc3NzY3IiwiaWF0IjoxNzE1NzUzNzQzfQ.ml-vMvWq5X8_FdILT9YIPv0oPc9Wlvj3f_N4VhHCAZA"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}